+++ frontend/app/src/state/atoms/schema.atom.ts
@@ -1,7 +1,4 @@
-import { queryClient } from "@/api/client";
 import { components } from "@/infraops";
-import { menuQueryOptions } from "@/screens/layout/menu-navigation/get-menu";
-import { MenuItem } from "@/screens/layout/menu-navigation/types";
 import { atom } from "jotai";
 
 export type iNodeSchema = components["schemas"]["APINodeSchema"];
@@ -22,23 +19,3 @@ export type iNamespace = {
 export const namespacesState = atom<iNamespace[]>([]);
 
 export const currentSchemaHashAtom = atom<string | null>(null);
-
-export const menuFlatAtom = atom((get) => {
-  const menuData = queryClient.getQueryData(menuQueryOptions().queryKey);
-  if (!menuData) return [];
-
-  const menuItems: MenuItem[] = [];
-
-  const flattenMenuItems = (menuItem: MenuItem) => {
-    if (menuItem.path !== "") menuItems.push(menuItem);
-
-    if (menuItem.children && menuItem.children.length > 0) {
-      menuItem.children.forEach(flattenMenuItems);
-    }
-  };
-
-  menuData.sections.object.forEach(flattenMenuItems);
-  menuData.sections.internal.forEach(flattenMenuItems);
-
-  return menuItems;
-});
